monolith_python/requirements.txt
absl-py==2.1.0
aerich==0.7.2
aiofiles==23.2.1
aiosqlite==0.17.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
dictdiffer==0.9.0
dm-tree==0.1.8
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.110.0
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
humanize==4.9.0
idna==3.6
iso8601==1.1.0
kagglehub==0.2.0
keras==2.15.0
keras-core==0.1.7
keras-nlp==0.8.2
Keras-Preprocessing==1.1.2
kombu==5.3.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.7
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
passlib==1.7.4
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pymystem3==0.2.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-text==2.15.0
termcolor==2.4.0
tf-keras==2.15.0
tomlkit==0.12.4
tornado==6.4
tortoise-orm==0.20.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
